Исправлены замечания по код-стайлу.
