+++ src/main/kotlin/dsx/bps/core/datamodel/Currency.kt
@@ -7,17 +7,17 @@ import dsx.bps.config.currencies.TrxConfig
 import dsx.bps.config.currencies.XrpConfig
 
 enum class Currency {
-    BTC{
+    BTC {
         override val coinConfigSpec = BtcConfig
     },
-    XRP{
+    XRP {
         override val coinConfigSpec = XrpConfig
     },
-    TRX{
+    TRX {
         override val coinConfigSpec = TrxConfig
     },
 
-    ETH{
+    ETH {
         override val coinConfigSpec = EthConfig
     };
