+++ src/main/kotlin/dsx/bps/crypto/common/Coin.kt
@@ -4,6 +4,7 @@ import com.uchuhimo.konf.Config
 import dsx.bps.core.datamodel.Currency
 import dsx.bps.core.datamodel.Tx
 import dsx.bps.core.datamodel.TxId
+import dsx.bps.crypto.eth.Connector
 import io.reactivex.subjects.PublishSubject
 import java.math.BigDecimal
 
@@ -12,7 +13,7 @@ abstract class Coin {
     abstract val currency: Currency
     abstract val config: Config
 
-    protected abstract val rpc: Any
+    protected abstract val rpc: Connector
     protected abstract val explorer: Explorer
 
     fun getTxEmitter(): PublishSubject<Tx> = explorer.emitter
