+++ src/main/kotlin/dsx/bps/crypto/eth/Connector.kt
@@ -1,0 +1,4 @@
+package dsx.bps.crypto.eth
+
+interface Connector {
+}
